pyproject.toml
[project]
name = "ucm-tfm-grupo-4"
version = "0.1.0"
description = "A NL-2-SQL system."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb==1.0.15",
    "google-genai>=1.30.0",
    "gradio>=3.36.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.29",
    "langgraph>=0.6.6",
    "notebook>=7.4.4",
    "openai==1.93.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pathlib>=1.0.1",
    "plotly>=6.3.0",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "regex>=2025.7.34",
    "setuptools<81",
    "sqlglot>=27.11.0",
]
